Lists languages from Google Translate API

Returns a list of supported languages for translation.

gl_translate_languages(target = "en")

Arguments

target
If specified, language names are localized in target language

Value

A tibble of supported languages

Details

Supported language codes, generally consisting of its ISO 639-1 identifier. (E.g. 'en', 'ja'). In certain cases, BCP-47 codes including language + region identifiers are returned (e.g. 'zh-TW', 'zh-CH')

See also

https://cloud.google.com/translate/docs/reference/languages

Other translations: gl_translate_detect, gl_translate

Examples

# NOT RUN {
# default english names of languages supported
gl_translate_languages()

# specify a language code to get other names, such as Danish
gl_translate_languages("da")

# }