googleLanguageR

This package contains functions for analysing language through the Google Cloud Machine Learning APIs

Details

For examples and documentation see the vignettes and the website: http://code.markedmondson.me/googleLanguageR/

See also

https://cloud.google.com/products/machine-learning/