Perform Natural Language Analysis on text

Analyse text entites, sentiment, and syntax using the Google Natural Language API

gl_nlp(string, nlp_type = c("annotateText", "analyzeEntities",
  "analyzeSentiment", "analyzeSyntax", "analyzeEntitySentiment"),
  type = c("PLAIN_TEXT", "HTML"), language = c("en", "zh", "zh-Hant", "fr",
  "de", "it", "ja", "ko", "pt", "es"), encodingType = c("UTF8", "UTF16",
  "UTF32", "NONE"), version = c("v1", "v1beta2", "v1beta1"))

Arguments

string
A character vector of text to detect language for, or Google Cloud Storage URIs

nlp_type
The type of Natural Language Analysis to perform. The default annotateText will perform all features in one call.

type
Whether input text is plain text or a HTML page

language
Language of source, must be supported by API.

encodingType
Text encoding that the caller uses to process the output

version
the API version

Value

A list of analysis of the text

Details

string can be a character vector, or a location of a file content on Google cloud Storage. This URI must be of the form gs://bucket_name/object_name

Encoding type can usually be left at default UTF8. See here for details: https://cloud.google.com/natural-language/docs/reference/rest/v1/EncodingType

The current language support is available here https://cloud.google.com/natural-language/docs/languages

See also

https://cloud.google.com/natural-language/docs/reference/rest/v1/documents

Examples

not_run({

  text <- "to administer medicince to animals is frequently a very difficult matter,
    and yet sometimes it's necessary to do so"
  nlp_result <- gl_nlp(text)

  head(nlp_result$tokens$text)

  head(nlp_result$tokens$partOfSpeech$tag)

  nlp_result$entities

  nlp_result$documentSentiment

})