Call Google Speech API

Turn audio into text

gl_speech(audio_source, encoding = c("LINEAR16", "FLAC", "MULAW", "AMR",
  "AMR_WB", "OGG_OPUS", "SPEEX_WITH_HEADER_BYTE"), sampleRateHertz = 16000L,
  languageCode = "en-US", maxAlternatives = 1L, profanityFilter = FALSE,
  speechContexts = NULL, asynch = FALSE)

Arguments

audio_source
File location of audio data, or Google Cloud Storage URI

encoding
Encoding of audio data sent

sampleRateHertz
Sample rate in Hertz of audio data. Valid values 8000-48000. Optimal 16000

languageCode
Language of the supplied audio as a BCP-47 language tag

maxAlternatives
Maximum number of recognition hypotheses to be returned. 0-30

profanityFilter
If TRUE will attempt to filter out profanities

speechContexts
An optional character vector of context to assist the speech recognition

asynch
If your audio_source is greater than 60 seconds, set this to TRUE to return an asynchronous call

Value

A tibble of three columns: transcript, the confidence with the number of rows equal to the maxAlternatives; a list-column of word timestamps. If asynch is TRUE, then an operation you will need to pass to gl_speech_op to get the finished result.

Details

Google Cloud Speech API enables developers to convert audio to text by applying powerful neural network models in an easy to use API. The API recognizes over 80 languages and variants, to support your global user base. You can transcribe the text of users dictating to an application’s microphone, enable command-and-control through voice, or transcribe audio files, among many other use cases. Recognize audio uploaded in the request, and integrate with your audio storage on Google Cloud Storage, by using the same technology Google uses to power its own products.

AudioEncoding

Audio encoding of the data sent in the audio message. All encodings support only 1 channel (mono) audio. Only FLAC and WAV include a header that describes the bytes of audio that follow the header. The other encodings are raw audio bytes with no header. For best results, the audio source should be captured and transmitted using a lossless encoding (FLAC or LINEAR16). Recognition accuracy may be reduced if lossy codecs, which include the other codecs listed in this section, are used to capture or transmit the audio, particularly if background noise is present.

Read more on audio encodings here https://cloud.google.com/speech/docs/encoding

WordInfo

Use tidyr::unnest() to extract the word columns if needed. startTime - Time offset relative to the beginning of the audio, and corresponding to the start of the spoken word. endTime - Time offset relative to the beginning of the audio, and corresponding to the end of the spoken word. word - The word corresponding to this set of information.

See also

https://cloud.google.com/speech/reference/rest/v1/speech/recognize

Examples

# NOT RUN {
test_audio <- system.file("woman1_wb.wav", package = "googleLanguageR")
result <- gl_speech(test_audio)

result2 <- gl_speech(test_audio, maxAlternatives = 2L)

result_brit <- gl_speech(test_audio, languageCode = "en-GB")

## extract word timestamps
tidyr::unnest(result_brit)

## make an asynchronous API request (mandatory for sound files over 60 seconds)
asynch <- gl_speech(test_audio, asynch = TRUE)

## Send to gl_speech_op() for status or finished result
gl_speech_op(asynch)

# }